refactor: 이벤트 발행 시점 변경
- 수용 여부 확인 -> 입장 처리 스케줄러
