+++ backend/queue-service/src/main/java/com/bidket/queue/application/service/QueueTrafficService.java
@@ -23,6 +23,7 @@ import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 import reactor.core.publisher.Sinks;
 
+import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
@@ -35,23 +36,9 @@ public class QueueTrafficService {
     private final QueueTrafficRepository trafficRepository;
     private final QueueManagementRepository managementRepository;
     private final TokenProvider tokenProvider;
-    private final ReactiveKafkaProducerTemplate<String, NotificationEvent> kafkaTemplate;
-    private final Sinks.Many<NotificationEvent> eventSink;
 
     @Value("${heartbeat.frequency}")
     private Long heartbeatFrequency;
-    private String topic;
-    private String key;
-
-    @PostConstruct
-    public void init() {
-        eventSink.asFlux()
-                .flatMap(event -> kafkaTemplate.send(topic, event))
-                .doOnComplete(() -> log.info("complete"))
-                .doOnError(e -> log.error("알림 이벤트 발행 실패: {}", e.getMessage(), e))
-                .onErrorResume(e -> Mono.empty())
-                .subscribe();
-    }
 
     @CheckQueueConfig
     public Mono<QueueEnterResponse> enterQueue(UUID userId, UUID auctionId) {
@@ -76,12 +63,6 @@ public class QueueTrafficService {
                 .flatMap(token -> {
                     if (!tokenProvider.validateToken(token, userId, auctionId))
                         return Mono.error(new QueueException(QueueErrorCode.INVALID_TOKEN));
-
-                    NotificationEvent event = NotificationEvent.builder()
-                            .auctionId(auctionId)
-                            .userId(userId)
-                            .enterTime(LocalDateTime.now())
-                            .build();
 
                     return Mono.just(QueueAccommodatableResponse.builder()
                             .auctionId(auctionId)
