+++ backend/queue-service/src/main/resources/application-local.yml
@@ -8,7 +8,7 @@ spring:
     import: optional:file:.env[.properties]
 
   datasource:
-    url: jdbc:postgresql://${POSTGRES_HOST:localhost}:${POSTGRES_PORT}/${USER_DB}
+    url: jdbc:postgresql://localhost:${POSTGRES_PORT}/${QUEUE_DB}
     username: ${POSTGRES_USER}
     password: ${POSTGRES_PASSWORD}
